Re-organizing render classes
Moved some render utilities out of model/util as they're not strictly
tied to models.
Deleted unnecessary data interfaces.
