+++ base/src/main/java/com/avogine/render/font/data/FontNameInfo.java
@@ -1,4 +1,4 @@
-package com.avogine.render.font;
+package com.avogine.render.font.data;
 
 /**
  * @param family
