--- base/src/main/java/com/avogine/render/model/mesh/data/VertexData.java
@@ -1,11 +1,0 @@
-package com.avogine.render.model.mesh.data;
-
-/**
- *
- */
-public interface VertexData {
-	/**
-	 * @return the number of vertices.
-	 */
-	public int getVertexCount();
-}
