+++ base/src/main/java/com/avogine/render/opengl/font/util/FontCache.java
@@ -1,4 +1,4 @@
-package com.avogine.render.font;
+package com.avogine.render.opengl.font.util;
 
 import java.util.*;
