+++ base/src/main/java/com/avogine/render/opengl/model/util/ModelLoader.java
@@ -14,14 +14,13 @@ import org.lwjgl.assimp.*;
 
 import com.avogine.logging.AvoLog;
 import com.avogine.render.model.animation.*;
-import com.avogine.render.model.mesh.data.VertexBuffers;
-import com.avogine.render.model.util.AssimpFileUtils;
+import com.avogine.render.model.mesh.data.*;
 import com.avogine.render.opengl.image.util.TextureCache;
 import com.avogine.render.opengl.model.Model;
 import com.avogine.render.opengl.model.material.*;
 import com.avogine.render.opengl.model.material.data.BlinnPhongData;
 import com.avogine.render.opengl.model.mesh.*;
-import com.avogine.render.opengl.model.mesh.data.MeshData;
+import com.avogine.render.util.AssimpFileUtils;
 
 /**
  *
