--- base/src/main/java/com/avogine/render/opengl/particle/data/ParticleVertexData.java
@@ -1,18 +1,0 @@
-package com.avogine.render.opengl.particle.data;
-
-import java.nio.FloatBuffer;
-
-import com.avogine.render.model.mesh.data.VertexData;
-
-/**
- * @param positions 
- * @param maxInstances 
- */
-public record ParticleVertexData(FloatBuffer positions, int maxInstances) implements VertexData {
-
-	@Override
-	public int getVertexCount() {
-		return positions.limit() / 3;
-	}
-
-}
