+++ base/src/main/java/com/avogine/render/opengl/ui/TextRender.java
@@ -9,9 +9,9 @@ import org.joml.Matrix4f;
 import org.lwjgl.system.*;
 
 import com.avogine.logging.AvoLog;
-import com.avogine.render.font.FontCache;
 import com.avogine.render.opengl.VAO;
 import com.avogine.render.opengl.font.Font;
+import com.avogine.render.opengl.font.util.FontCache;
 import com.avogine.render.opengl.shader.FontShader;
 import com.avogine.render.opengl.ui.text.TextMesh;
 import com.avogine.util.resource.ResourceConstants;
