+++ base/src/main/java/com/avogine/render/util/AssimpFileUtils.java
@@ -1,4 +1,4 @@
-package com.avogine.render.model.util;
+package com.avogine.render.util;
 
 import static org.lwjgl.system.MemoryUtil.*;
