+++ base/src/main/java/com/avogine/render/util/BuildFunction.java
@@ -1,17 +1,18 @@
-package com.avogine.render.model.util;
+package com.avogine.render.util;
 
 import org.lwjgl.util.par.ParShapesMesh;
 
 /**
  * @param <T> 
+ *
  */
-public interface InstancedBuildFunction<T> {
+@FunctionalInterface
+public interface BuildFunction<T> {
 
 	/**
 	 * @param parShapesMesh
-	 * @param instanceCount
 	 * @return a built instance of {@code T} from the given {@link ParShapesMesh} structure.
 	 */
-	public T build(ParShapesMesh parShapesMesh, int instanceCount);
+	public T build(ParShapesMesh parShapesMesh);
 	
 }
