+++ base/src/main/java/com/avogine/render/util/InstancedBuildFunction.java
@@ -1,18 +1,17 @@
-package com.avogine.render.model.util;
+package com.avogine.render.util;
 
 import org.lwjgl.util.par.ParShapesMesh;
 
 /**
  * @param <T> 
- *
  */
-@FunctionalInterface
-public interface BuildFunction<T> {
+public interface InstancedBuildFunction<T> {
 
 	/**
 	 * @param parShapesMesh
+	 * @param instanceCount
 	 * @return a built instance of {@code T} from the given {@link ParShapesMesh} structure.
 	 */
-	public T build(ParShapesMesh parShapesMesh);
+	public T build(ParShapesMesh parShapesMesh, int instanceCount);
 	
 }
